Fix broken responses for $return bindings
When authenticating HTTP trigger functions that use the $return binding
for the HTTP output instead of a name such as "res", assigning a
response to context.res doesn't work because the Azure Functions runtime
is looking for the res object in the function return value. This commit
fixes that issue in addition to several other code structure
improvements.